+++ packages/babel-plugin-lingui-macro/src/constants.ts
@@ -1,7 +1,4 @@
 export const EXTRACT_MARK = "i18n"
-export const MACRO_LEGACY_PACKAGE = "@lingui/macro"
-export const MACRO_CORE_PACKAGE = "@lingui/core/macro"
-export const MACRO_REACT_PACKAGE = "@lingui/react/macro"
 
 export enum MsgDescriptorPropKey {
   id = "id",
@@ -21,6 +18,7 @@ export enum JsMacroName {
   defineMessage = "defineMessage",
   arg = "arg",
   useLingui = "useLingui",
+  ph = "ph",
 }
 
 export enum JsxMacroName {