+++ packages/vite-plugin/src/index.ts
@@ -4,31 +4,48 @@ import {
   getCatalogs,
   getCatalogForFile,
   getCatalogDependentFiles,
+  createMissingErrorMessage,
+  createCompilationErrorMessage,
 } from "@lingui/cli/api"
 import path from "path"
 import type { Plugin } from "vite"
 
 const fileRegex = /(\.po|\?lingui)$/
 
-type LinguiConfigOpts = {
+export type LinguiPluginOpts = {
   cwd?: string
   configPath?: string
   skipValidation?: boolean
+
+  /**
+   * If true would fail compilation on missing translations
+   **/
+  failOnMissing?: boolean
+
+  /**
+   * If true would fail compilation on message compilation errors
+   **/
+  failOnCompileError?: boolean
 }
 
-export function lingui(linguiConfig: LinguiConfigOpts = {}): Plugin[] {
+export function lingui({
+  failOnMissing,
+  failOnCompileError,
+  ...linguiConfig
+}: LinguiPluginOpts = {}): Plugin[] {
   const config = getConfig(linguiConfig)
+
+  const macroIds = new Set([
+    ...config.macro.corePackage,
+    ...config.macro.jsxPackage,
+  ])
 
   return [
     {
       name: "vite-plugin-lingui-report-macro-error",
       enforce: "pre",
       resolveId(id) {
-        if (
-          id.includes("@lingui/macro") ||
-          id.includes("@lingui/core/macro") ||
-          id.includes("@lingui/react/macro")
-        ) {
+        if (macroIds.has(id)) {
           throw new Error(
             `The macro you imported from "${id}" is being executed outside the context of compilation. \n` +
               `This indicates that you don't configured correctly one of the "babel-plugin-macros" / "@lingui/swc-plugin" / "babel-plugin-lingui-macro"` +
@@ -46,9 +63,10 @@ export function lingui(linguiConfig: LinguiConfigOpts = {}): Plugin[] {
           config.optimizeDeps = {}
         }
         config.optimizeDeps.exclude = config.optimizeDeps.exclude || []
-        config.optimizeDeps.exclude.push("@lingui/macro")
-        config.optimizeDeps.exclude.push("@lingui/core/macro")
-        config.optimizeDeps.exclude.push("@lingui/react/macro")
+
+        for (const macroId of macroIds) {
+          config.optimizeDeps.exclude.push(macroId)
+        }
       },
       async transform(src, id) {
         if (fileRegex.test(id)) {
@@ -78,19 +96,54 @@ Please check that catalogs.path is filled properly.\n`
           const dependency = await getCatalogDependentFiles(catalog, locale)
           dependency.forEach((file) => this.addWatchFile(file))
 
-          const messages = await catalog.getTranslations(locale, {
-            fallbackLocales: config.fallbackLocales,
-            sourceLocale: config.sourceLocale,
-          })
+          const { messages, missing: missingMessages } =
+            await catalog.getTranslations(locale, {
+              fallbackLocales: config.fallbackLocales,
+              sourceLocale: config.sourceLocale,
+            })
 
-          const compiled = createCompiledCatalog(locale, messages, {
-            strict: false,
-            namespace: "es",
-            pseudoLocale: config.pseudoLocale,
-          })
+          if (
+            failOnMissing &&
+            locale !== config.pseudoLocale &&
+            missingMessages.length > 0
+          ) {
+            const message = createMissingErrorMessage(
+              locale,
+              missingMessages,
+              "loader"
+            )
+            throw new Error(
+              `${message}\nYou see this error because \`failOnMissing=true\` in Vite Plugin configuration.`
+            )
+          }
+
+          const { source: code, errors } = createCompiledCatalog(
+            locale,
+            messages,
+            {
+              namespace: "es",
+              pseudoLocale: config.pseudoLocale,
+            }
+          )
+
+          if (errors.length) {
+            const message = createCompilationErrorMessage(locale, errors)
+
+            if (failOnCompileError) {
+              throw new Error(
+                message +
+                  `These errors fail build because \`failOnCompileError=true\` in Lingui Vite plugin configuration.`
+              )
+            } else {
+              console.warn(
+                message +
+                  `You can fail the build on these errors by setting \`failOnCompileError=true\` in Lingui Vite Plugin configuration.`
+              )
+            }
+          }
 
           return {
-            code: compiled,
+            code,
             map: null, // provide source map if available
           }
         }